add CI, tests and support for node>0.12.15, close #2 and #14
