--- example/file1.md
@@ -1,17 +1,0 @@
-# Example File
-
-In this file I will include the second file twice:
-
-{{ Include file2.md }}
-
-
-
-{{ INCLUDE file2.md }}
-
-This is the end of the full document
-
-{{ message }}
-
-{{
-    message
-    }}
