--- example/file2.md
@@ -1,15 +1,0 @@
-This is the second file (cool uh?). 
-
-Now a custom command loop:
-
-{{loop(Inside loop,10)}}
-
-My Readme File:
-{{include ../README.md}}
-
-And a code file:
-
-```js
-{{include ../package.json}}
-```
-_package.json_
