+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "xejs",
-  "version": "0.3.0",
+  "version": "0.3.1",
   "description": "Recursive templating. What could go wrong?",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "istanbul cover _mocha"
   },
   "repository": {
     "type": "git",
@@ -18,5 +18,13 @@
   "homepage": "https://github.com/angrykoala/xejs#readme",
   "dependencies": {
     "ejs": "^2.5.2"
+  },
+  "devDependencies": {
+    "chai": "^3.5.0",
+    "istanbul": "^0.4.5",
+    "mocha": "^3.0.2"
+  },
+  "engines": {
+    "node": ">= 0.12.15"
   }
 }
