+++ test/file2.md
@@ -1,0 +1,3 @@
+## Second file
+
+Second file content
